[#67444] ArrayMemory: Fix size check
Passing negative values makes no sense and with `ulong` we can properly
check the value and provide in exception.

Previously, for example for 0x80000000, there was an exception that size
was -2147483648.

+++ src/Emulator/Peripherals/Peripherals/CAN/S32K3XX_FlexCAN/S32K3XX_FlexCAN.cs
@@ -39,7 +39,7 @@ namespace Antmicro.Renode.Peripherals.CAN
             IRQ = new GPIO();
 
             messageBufferRange = new Range((ulong)Registers.MessageBuffer, numberOfMessageBuffers * 8);
-            messageBuffers = new ArrayMemory((int)messageBufferRange.Size);
+            messageBuffers = new ArrayMemory(messageBufferRange.Size);
             messageBufferInterruptEnable = new IFlagRegisterField[numberOfMessageBuffers];
             messageBufferInterrupt = new IFlagRegisterField[numberOfMessageBuffers];
             messageBufferSize = new IEnumRegisterField<MessageBufferSize>[MessageBufferRegionsCount];
